Cleaned gui frame
